src/defense/BackdoorDefense/requirements.txt
hydra-core==1.3.2
inflection==0.5.1
language_tool_python==2.8
matplotlib==3.5.3
nltk==3.8.1
# numpy==1.21.6
OpenHowNet==2.0
pandas==1.1.5
pyinflect==0.5.1
pytorch_lightning==1.9.5
scikit_learn==1.0.2
scipy==1.7.3
seaborn==0.13.2
sentence_transformers==2.2.2
statsmodels==0.13.5
tensorflow==2.19.0
torch==1.13.1
tqdm==4.66.1
transformers==4.30.2
umap==0.1.1
sentencepiece